requirements.txt
asgiref
coverage
Django
djangorestframework
psycopg2-binary
sqlparse
